+++ classes/Generator/Type/Clone.php
@@ -1,0 +1,11 @@
+<?php defined('SYSPATH') OR die('No direct script access.');
+/**
+ * Generator Clone type.
+ *
+ * @package    Generator
+ * @category   Generator/Types
+ * @author     Zeebee
+ * @copyright  (c) 2012 Zeebee
+ * @license    BSD revised
+ */
+class Generator_Type_Clone extends Generator_Generator_Type_Clone {}