+++ views/generator/type_clone.php
@@ -1,5 +1,5 @@
 /**
- * Class <?php echo $name ?>.
+ * <?php echo ucfirst($type).' '.$name ?>, cloned from <?php echo $source ?>.
  *
  * @package    <?php echo $package ?> 
  * @category   <?php echo $category ?> 
@@ -8,31 +8,23 @@
  * @license    <?php echo $license ?> 
  */
 <?php 
-	echo ( ! empty($abstract) ? 'abstract class ' : 'class ');
-	echo $name;	
+	echo ( ! empty($modifiers) ? ($modifiers.' '.$type) : $type);
+	echo ' '.$name;	
 	if ( ! empty($extends)) {echo ' extends '.$extends;} 
 	if ( ! empty($implements)) {echo ' implements '.$implements;}
-	if ( ! empty($blank)) {echo ' {}';} else { ?> 
+	if ( ! empty($blank)): echo ' {}'; else : ?> 
 {
-	/**
-	 * @var  string  some string
-	 */
-	public $some_string;
+<?php if ( ! empty($constants)) foreach ($constants as $constant => $c): ?>
+	<?php echo $c['comment'] ?> 
+	<?php echo $c['declaration'] ?>;
 
-	/**
-	 * Short description.
-	 *
-	 * Long method description.
-	 *
-	 * @param  string  $param  Some string
-	 * @return void
-	 */
-	public function some_method($param)
-	{
-		// Method implementation
-	}
-
+<?php endforeach; ?>
 <?php foreach (array('static', 'public', 'abstract', 'other') as $group): ?>
+<?php	if (isset($properties[$group])) foreach ($properties[$group] as $property => $p): ?>
+	<?php echo $p['doccomment'] ?> 
+	<?php echo $p['declaration'] ?>; 
+
+<?php endforeach; ?>
 <?php	if (isset($methods[$group])) foreach ($methods[$group] as $method => $m): ?>
 	<?php echo $m['doccomment'] ?> 
 	<?php echo $m['signature'] ?>
@@ -45,4 +37,4 @@
 <?php endforeach; ?>
 <?php endforeach; ?>
 } // End <?php echo $name ?>
-<?php } ?> 
+<?php endif; ?>